requirements.txt
# Python dependencies for GitLab Security Audit System
#
# Core security scanning tools
semgrep>=1.45.0

# HTTP client for API interactions and downloads
requests>=2.31.0

# CLI framework for building command-line interfaces
click>=8.1.7

# Cross-platform colored terminal output
colorama>=0.4.6

# Pretty-print tabular data
tabulate>=0.9.0

# Environment variable management
python-dotenv>=1.0.0

# JSON manipulation and validation
jsonschema>=4.17.0

# YAML parsing (for configuration files)
PyYAML>=6.0

# Date/time handling
python-dateutil>=2.8.2